+++ src/components/Editor/Project/Project.js
@@ -1,19 +1,22 @@
-import './Project.scss';
-
+/* eslint-disable react-hooks/exhaustive-deps */
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector} from 'react-redux'
-
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
 import 'react-tabs/style/react-tabs.css'
 import 'react-toastify/dist/ReactToastify.css'
 
+import './Project.scss';
 import EditorPanel from '../EditorPanel/EditorPanel'
 import FilePane from '../../FilePane/FilePane'
 import Output from '../Output/Output'
 import RenameFile from '../../Modals/RenameFile'
 import RunnerControls from '../../RunButton/RunnerControls'
-import { syncProject } from '../EditorSlice';
+import { expireJustLoaded, setHasShownSavePrompt, syncProject } from '../EditorSlice';
 import { isOwner } from '../../../utils/projectHelpers'
+import NotFoundModal from '../../Modals/NotFoundModal';
+import AccessDeniedNoAuthModal from '../../Modals/AccessDeniedNoAuthModal';
+import AccessDeniedWithAuthModal from '../../Modals/AccessDeniedWithAuthModal';
+import { showLoginPrompt, showSavePrompt } from '../../../utils/Notifications';
 
 const Project = (props) => {
   const dispatch = useDispatch()
@@ -22,6 +25,11 @@ const Project = (props) => {
   const project = useSelector((state) => state.editor.project)
   const modals = useSelector((state) => state.editor.modals)
   const renameFileModalShowing = useSelector((state) => state.editor.renameFileModalShowing)
+  const notFoundModalShowing = useSelector((state) => state.editor.notFoundModalShowing)
+  const accessDeniedNoAuthModalShowing = useSelector((state) => state.editor.accessDeniedNoAuthModalShowing)
+  const accessDeniedWithAuthModalShowing = useSelector((state) => state.editor.accessDeniedWithAuthModalShowing)
+  const justLoaded = useSelector((state) => state.editor.justLoaded)
+  const hasShownSavePrompt = useSelector((state) => state.editor.hasShownSavePrompt)
 
   useEffect(() => {
     if (forWebComponent) {
@@ -34,6 +42,12 @@ const Project = (props) => {
       }
       else {
         localStorage.setItem(project.identifier || 'project', JSON.stringify(project))
+        if (justLoaded) {
+          dispatch(expireJustLoaded())
+        } else if (!hasShownSavePrompt) {
+          user ? showSavePrompt() : showLoginPrompt()
+          dispatch(setHasShownSavePrompt())
+        }
       }
     }, 2000);
      
@@ -64,6 +78,9 @@ const Project = (props) => {
         <Output />
       </div>
       {(renameFileModalShowing && modals.renameFile) ? <RenameFile /> : null}
+      {(notFoundModalShowing) ? <NotFoundModal /> : null}
+      {(accessDeniedNoAuthModalShowing) ? <AccessDeniedNoAuthModal /> : null}
+      {(accessDeniedWithAuthModalShowing) ? <AccessDeniedWithAuthModal /> : null}
     </div>
   )
 };