+++ src/utils/Notifications.test.js
@@ -1,5 +1,5 @@
 import { toast } from 'react-toastify'
-import { showRemixedMessage, showSavedMessage } from "./Notifications";
+import { showLoginPrompt, showRemixedMessage, showSavedMessage, showSavePrompt } from "./Notifications";
 
 jest.mock('../i18n', () => ({
   t: (string) => string
@@ -14,4 +14,14 @@ test('Calling showRemixedMessage calls toast with correct string', () => {
 test('Calling showSavedMessage calls toast with correct string', () => {
   showSavedMessage()
   expect(toast).toHaveBeenCalledWith('notifications.projectSaved', expect.anything())
+})
+
+test('Calling showSavePrompt calls toast with correct string', () => {
+  showSavePrompt()
+  expect(toast).toHaveBeenCalledWith('notifications.savePrompt', expect.anything())
+})
+
+test('Calling showLoginPrompt calls toast with correct string', () => {
+  showLoginPrompt()
+  expect(toast).toHaveBeenCalledWith('notifications.loginPrompt', expect.anything())
 })