+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Plausbile event tracking for login, remix, save and code run (#250)
+- Message prompting users to login or save if they make non-autosaved changes (#291)
+- Unit tests for the autosave trigger (#291)
+- Project not found and access denied modals shown on project loading error (#298)
+
+## Changed
+
+- Bump terser from 4.8.0 to 4.8.1 (#143)
+- Updated buttons and tabs styling (#296)
+- Updated icons (#296)
+
 ### Fixed
 
 - Moved `FileMenu` click handler to `MenuItem` for better keyboard support (#300)